Better Siderbar
Ran into the problem of the sidebar having too many list items, which caused the page to be longer than it needed to be.

Fixed this by creating a max-h-sidebar variable in the tailwind config. I tried to keep the max-h in between the header and footer at first, but decided to only worry about keeping it under the header.
This means that every page inside a layout.ts file with a sidebar (with a long list) will have a body scrollbar to get to the footer, but it's the lesser evil. It'll actually help prevent some layout shifting that occurs when navigating around.

• Defined the header to have a height of 50px.
• Sticky positioned the <aside>'s first child <div> to be 50px from the top. This <div> houses the lists of navigation.
• The trick was to set these class names together in that <div>: "h-min max-h-sidebar".
• The sidebar maintains a full height now as well, making the open/close transition nicer.